+++ tests/testthat/indention_multiple/if_else_curly-in.R
@@ -37,3 +37,18 @@ foo <- function(x) {
     2
   }
 }
+
+
+if (TRUE) {
+  3
+} else
+  if (FALSE) {
+  4
+}
+
+if (TRUE) {
+  3
+} else # comment
+  if (FALSE) {
+    4
+  }