+++ R/zzz.R
@@ -5,6 +5,7 @@
     styler.cache_root = NULL,
     styler.cache_name = styler_version,
     styler.colored_print.vertical = TRUE,
+    styler.ignore_alignment = FALSE,
     styler.ignore_start = "styler: off",
     styler.ignore_stop = "styler: on",
     styler.quiet = FALSE,