+++ tests/testthat/indention_multiple/if_else_curly-out.R
@@ -36,3 +36,17 @@ foo <- function(x) {
     2
   }
 }
+
+
+if (TRUE) {
+  3
+} else  if (FALSE) {
+  4
+}
+
+if (TRUE) {
+  3
+} else # comment
+if (FALSE) {
+  4
+}