+++ NEWS.md
@@ -1,10 +1,19 @@
 # styler 1.7.0.9000 (Development version)
 
-* `filetype` `.qmd` is now supported, but not turned on by default (#931).
+**Features**
 
+* `filetype` `.qmd` is now supported, but not turned on by default (#931).
+* new R option `styler.ignore_alignment` controls if alignment should be 
+  detected (and preserved) or not (#932).
+  
+**Bug Fixes**
+
+* the cache is also invalidated on changing the stylerignore markers (#932).
 
 
 # styler 1.7.0
+
+* if `else` follows directly after `if`, line breaks are removed (#935).
 
 **API changes**